+++ test/babel-register.js
@@ -1,4 +1,0 @@
-// Disable the babel cache
-require('@babel/register')({
-    cache: false,
-});