+++ test/fixtures/non-json-import/entry.js
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/no-webpack-loader-syntax
-import theoContent from '../../../src/index.js!hson-loader!./props.hson';
-
-export default theoContent;