+++ test/fixtures/non-nested-imports/entry-query-options.js
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/no-webpack-loader-syntax
-import theoContent from '../../../src/index.js?{"format":{"type":"raw.json"},"propToDelete":"three"}!./props.json';
-
-export default theoContent;